Shrink FuncState::pinnedreg to short since we only need 9 bits

+++ src/lparser.h
@@ -357,7 +357,7 @@ typedef struct FuncState {
   lu_byte freereg;  /* first free register */
   lu_byte iwthabs;  /* instructions issued since last absolute line info */
   lu_byte needclose;  /* function needs to close upvalues when returning */
-  int pinnedreg;
+  short pinnedreg;  /* [Pluto] index of register that may not be free'd or -1 */
 } FuncState;
